requirements.txt
flask==3.0.0
flask-restx==1.3.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
lxml==4.9.3
flask-cors==4.0.0
python-dotenv==1.0.0
streamlit==1.28.1
plotly==5.17.0
PyJWT==2.8.0
bcrypt==4.1.2
scikit-learn==1.3.2
numpy==1.24.3
joblib==1.3.2